requirements.txt
bs4
rich
requests
dearpygui
dnspython
cryptography